Support 202504
fixes #1145
fixes #1614
fixes #1639
fixes #1682
fixes #1705
fixes #1719
fixes #1732
fixes #1737
fixes #1757

+++ src/peeringdb_server/admin.py
@@ -34,7 +34,7 @@ from django.db import models, transaction
 from django.db.models import Q
 from django.db.utils import OperationalError
 from django.forms import DecimalField
-from django.http import HttpResponseForbidden, JsonResponse
+from django.http import HttpResponseForbidden, HttpResponseRedirect, JsonResponse
 from django.shortcuts import redirect
 from django.template import loader
 from django.template.response import TemplateResponse
@@ -465,6 +465,15 @@ def soft_delete(modeladmin, request, queryset):
             return
 
         for row in queryset:
+            if isinstance(row, IXLanPrefix) and not row.deletable:
+                messages.error(
+                    request,
+                    _("Protected object '{}': {}").format(
+                        row, row.not_deletable_reason
+                    ),
+                )
+                continue
+
             try:
                 row.delete()
             except ProtectedAction as err:
@@ -1674,6 +1683,34 @@ class IXLanPrefixAdmin(SoftDeleteAdmin, ISODateTimeMixin):
     def ix(self, obj):
         return obj.ixlan.ix
 
+    def change_view(self, request, object_id, form_url="", extra_context=None):
+        """
+        Show message to the admin page if the prefix is not deletable
+        """
+        obj = self.get_object(request, object_id)
+        response = super().change_view(request, object_id, form_url, extra_context)
+
+        if obj and not obj.deletable:
+            messages.warning(request, f"⚠️ {obj.not_deletable_reason}")
+
+        return response
+
+    def delete_model(self, request, obj):
+        """Check the deletable property"""
+        if not obj.deletable:
+            messages.error(request, obj.not_deletable_reason)
+            return
+        obj.delete()
+
+    def delete_view(self, request, object_id, extra_context=None):
+        """Custom delete view to check deletable property before confirming deletion"""
+        obj = self.get_object(request, object_id)
+        if obj and not obj.deletable:
+            messages.error(request, obj.not_deletable_reason)
+            # Redirect back to the changelist page
+            return HttpResponseRedirect("../")
+        return super().delete_view(request, object_id, extra_context)
+
 
 class NetworkIXLanAdminForm(StatusForm):
     net_side = baseForms.ChoiceField(required=False)
@@ -1781,7 +1818,7 @@ class NetworkContactAdmin(SoftDeleteAdmin, ISODateTimeMixin):
 class NetworkFacilityAdmin(SoftDeleteAdmin, ISODateTimeMixin):
     list_display = ("id", "net", "facility", "status", "iso_created", "iso_updated")
     search_fields = ("network__asn", "network__name", "facility__name")
-    readonly_fields = ("id", "net")
+    readonly_fields = ("id", "net", "local_asn")
     list_filter = (StatusFilter,)
     form = StatusForm
